Move CUDA interop API back into video
