--- include/nbl/ext/CUDAInterop/CUDAInterop.h
@@ -1,13 +1,0 @@
-// Copyright (C) 2018-2020 - DevSH Graphics Programming Sp. z O.O.
-// This file is part of the "Nabla Engine".
-// For conditions of distribution and use, see copyright notice in nabla.h
-#ifndef _NBL_EXT_CUDA_INTEROP_H_INCLUDED_
-#define _NBL_EXT_CUDA_INTEROP_H_INCLUDED_
-
-#include "nbl/ext/CUDAInterop/CCUDADevice.h"
-#include "nbl/ext/CUDAInterop/CCUDAExportableMemory.h"
-#include "nbl/ext/CUDAInterop/CCUDAHandler.h"
-#include "nbl/ext/CUDAInterop/CCUDAImportedMemory.h"
-#include "nbl/ext/CUDAInterop/CCUDAImportedSemaphore.h"
-
-#endif
