+++ include/nbl/video/CCUDAImportedMemory.h
@@ -1,5 +1,5 @@
-#ifndef _NBL_EXT_CUDA_INTEROP_C_CUDA_IMPORTED_MEMORY_H_
-#define _NBL_EXT_CUDA_INTEROP_C_CUDA_IMPORTED_MEMORY_H_
+#ifndef _NBL_VIDEO_C_CUDA_IMPORTED_MEMORY_H_
+#define _NBL_VIDEO_C_CUDA_IMPORTED_MEMORY_H_
 
 #include "nbl/video/declarations.h"
