+++ include/nbl/video/CUDAInterop.h
@@ -1,0 +1,13 @@
+// Copyright (C) 2018-2020 - DevSH Graphics Programming Sp. z O.O.
+// This file is part of the "Nabla Engine".
+// For conditions of distribution and use, see copyright notice in nabla.h
+#ifndef _NBL_VIDEO_CUDA_INTEROP_H_INCLUDED_
+#define _NBL_VIDEO_CUDA_INTEROP_H_INCLUDED_
+
+#include "nbl/video/CCUDADevice.h"
+#include "nbl/video/CCUDAExportableMemory.h"
+#include "nbl/video/CCUDAHandler.h"
+#include "nbl/video/CCUDAImportedMemory.h"
+#include "nbl/video/CCUDAImportedSemaphore.h"
+
+#endif
