+++ src/nbl/ext/CUDAInterop/smoke/clean_opt_in.cpp
@@ -1,4 +1,4 @@
-#include "nbl/ext/CUDAInterop/CUDAInterop.h"
+#include "nbl/video/CUDAInterop.h"
 #include "nbl/system/IApplicationFramework.h"
 
 #include <type_traits>
