+++ src/nbl/video/CCUDAHandler.cpp
@@ -2,7 +2,7 @@
 // This file is part of the "Nabla Engine".
 // For conditions of distribution and use, see copyright notice in nabla.h
 
-#include "nbl/ext/CUDAInterop/CUDAInterop.h"
+#include "nbl/video/CUDAInterop.h"
 #include "nbl/system/ModuleLookupUtils.h"
 
 #include "nlohmann/json.hpp"
