+++ src/nbl/video/CUDAInteropNativeState.hpp
@@ -1,5 +1,5 @@
-#ifndef _NBL_EXT_CUDA_INTEROP_NATIVE_STATE_H_INCLUDED_
-#define _NBL_EXT_CUDA_INTEROP_NATIVE_STATE_H_INCLUDED_
+#ifndef _NBL_VIDEO_CUDA_INTEROP_NATIVE_STATE_H_INCLUDED_
+#define _NBL_VIDEO_CUDA_INTEROP_NATIVE_STATE_H_INCLUDED_
 
 #include "nbl/ext/CUDAInterop/CUDAInteropNative.h"
